+++ options.go
@@ -30,10 +30,10 @@ import (
 )
 
 // Option contains configuration values for Connection
-type Option func(ctx context.Context, c *connection) error
+type Option func(ctx context.Context, c *Driver) error
 
 func WithStaticCredentials(user, password string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.userInfo = &dsn.UserInfo{
 			User:     user,
 			Password: password,
@@ -55,14 +55,14 @@ func WithAccessTokenCredentials(accessToken string) Option {
 
 // WithUserAgent add provided user agent value to all api requests
 func WithUserAgent(userAgent string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithUserAgent(userAgent))
 		return nil
 	}
 }
 
 func WithRequestsType(requestsType string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithRequestsType(requestsType))
 		return nil
 	}
@@ -76,7 +76,7 @@ func WithRequestsType(requestsType string) Option {
 //
 // (connection string will be required string param of ydb.Open)
 func WithConnectionString(connectionString string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		if connectionString == "" {
 			return nil
 		}
@@ -99,7 +99,7 @@ func WithConnectionString(connectionString string) Option {
 
 // WithConnectionTTL defines duration for parking idle connections
 func WithConnectionTTL(ttl time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithConnectionTTL(ttl))
 		return nil
 	}
@@ -111,7 +111,7 @@ func WithConnectionTTL(ttl time.Duration) Option {
 //
 // For making connection string from endpoint+database+secure - use sugar.DSN()
 func WithEndpoint(endpoint string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithEndpoint(endpoint))
 		return nil
 	}
@@ -123,7 +123,7 @@ func WithEndpoint(endpoint string) Option {
 //
 // For making connection string from endpoint+database+secure - use sugar.DSN()
 func WithDatabase(database string) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithDatabase(database))
 		return nil
 	}
@@ -135,7 +135,7 @@ func WithDatabase(database string) Option {
 //
 // For making connection string from endpoint+database+secure - use sugar.DSN()
 func WithSecure(secure bool) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithSecure(secure))
 		return nil
 	}
@@ -145,7 +145,7 @@ func WithSecure(secure bool) Option {
 //
 // Warning: WithInsecure lost current TLS config.
 func WithInsecure() Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithSecure(false))
 		return nil
 	}
@@ -153,7 +153,7 @@ func WithInsecure() Option {
 
 // WithMinTLSVersion set minimum TLS version acceptable for connections
 func WithMinTLSVersion(minVersion uint16) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithMinTLSVersion(minVersion))
 		return nil
 	}
@@ -161,7 +161,7 @@ func WithMinTLSVersion(minVersion uint16) Option {
 
 // WithTLSSInsecureSkipVerify applies InsecureSkipVerify flag to TLS config
 func WithTLSSInsecureSkipVerify() Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithTLSSInsecureSkipVerify())
 		return nil
 	}
@@ -201,7 +201,7 @@ func WithAnonymousCredentials() Option {
 
 // WithCreateCredentialsFunc add callback funcion to provide requests credentials
 func WithCreateCredentialsFunc(createCredentials func(ctx context.Context) (credentials.Credentials, error)) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		creds, err := createCredentials(ctx)
 		if err != nil {
 			return xerrors.WithStackTrace(err)
@@ -220,7 +220,7 @@ func WithCredentials(c credentials.Credentials) Option {
 }
 
 func WithBalancer(balancer *balancerConfig.Config) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithBalancer(balancer))
 		return nil
 	}
@@ -230,7 +230,7 @@ func WithBalancer(balancer *balancerConfig.Config) Option {
 //
 // Default dial timeout is config.DefaultDialTimeout
 func WithDialTimeout(timeout time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithDialTimeout(timeout))
 		return nil
 	}
@@ -240,7 +240,7 @@ func WithDialTimeout(timeout time.Duration) Option {
 //
 // This option does not replace collected option, instead it will append provided options.
 func With(options ...config.Option) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, options...)
 		return nil
 	}
@@ -248,7 +248,7 @@ func With(options ...config.Option) Option {
 
 // MergeOptions concatentaes provided options to one cumulative value.
 func MergeOptions(opts ...Option) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		for _, o := range opts {
 			if o != nil {
 				if err := o(ctx, c); err != nil {
@@ -262,7 +262,7 @@ func MergeOptions(opts ...Option) Option {
 
 // WithDiscoveryInterval sets interval between cluster discovery calls.
 func WithDiscoveryInterval(discoveryInterval time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.discoveryOptions = append(c.discoveryOptions, discoveryConfig.WithInterval(discoveryInterval))
 		return nil
 	}
@@ -270,7 +270,7 @@ func WithDiscoveryInterval(discoveryInterval time.Duration) Option {
 
 // WithTraceDriver returns deadline which has associated Driver with it.
 func WithTraceDriver(trace trace.Driver, opts ...trace.DriverComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithTrace(trace, opts...))
 		return nil
 	}
@@ -278,7 +278,7 @@ func WithTraceDriver(trace trace.Driver, opts ...trace.DriverComposeOption) Opti
 
 // WithCertificate appends certificate to TLS config root certificates
 func WithCertificate(cert *x509.Certificate) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithCertificate(cert))
 		return nil
 	}
@@ -297,7 +297,7 @@ func WithCertificatesFromFile(caFile string, opts ...certificates.FromFileOption
 	if file, err := filepath.EvalSymlinks(caFile); err == nil {
 		caFile = file
 	}
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		certs, err := certificates.FromFile(caFile, opts...)
 		if err != nil {
 			return xerrors.WithStackTrace(err)
@@ -316,7 +316,7 @@ func WithCertificatesFromFile(caFile string, opts ...certificates.FromFileOption
 // Warning: all early TLS config changes (such as WithCertificate, WithCertificatesFromFile, WithCertificatesFromPem,
 // WithMinTLSVersion, WithTLSSInsecureSkipVerify) will be lost
 func WithTLSConfig(tlsConfig *tls.Config) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.options = append(c.options, config.WithTLSConfig(tlsConfig))
 		return nil
 	}
@@ -324,7 +324,7 @@ func WithTLSConfig(tlsConfig *tls.Config) Option {
 
 // WithCertificatesFromPem appends certificates from pem-encoded data to TLS config root certificates
 func WithCertificatesFromPem(bytes []byte, opts ...certificates.FromPemOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		certs, err := certificates.FromPem(bytes, opts...)
 		if err != nil {
 			return xerrors.WithStackTrace(err)
@@ -339,7 +339,7 @@ func WithCertificatesFromPem(bytes []byte, opts ...certificates.FromPemOption) O
 // WithTableConfigOption collects additional configuration options for table.Client.
 // This option does not replace collected option, instead it will appen provided options.
 func WithTableConfigOption(option tableConfig.Option) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, option)
 		return nil
 	}
@@ -347,7 +347,7 @@ func WithTableConfigOption(option tableConfig.Option) Option {
 
 // WithSessionPoolSizeLimit set max size of internal sessions pool in table.Client
 func WithSessionPoolSizeLimit(sizeLimit int) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, tableConfig.WithSizeLimit(sizeLimit))
 		return nil
 	}
@@ -357,12 +357,12 @@ func WithSessionPoolSizeLimit(sizeLimit int) Option {
 //
 // Deprecated: table client do not supports background session keep-aliving now
 func WithSessionPoolKeepAliveMinSize(keepAliveMinSize int) Option {
-	return func(ctx context.Context, c *connection) error { return nil }
+	return func(ctx context.Context, c *Driver) error { return nil }
 }
 
 // WithSessionPoolIdleThreshold defines keep-alive interval for idle sessions
 func WithSessionPoolIdleThreshold(idleThreshold time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, tableConfig.WithIdleThreshold(idleThreshold))
 		c.databaseSQLOptions = append(
 			c.databaseSQLOptions,
@@ -374,12 +374,12 @@ func WithSessionPoolIdleThreshold(idleThreshold time.Duration) Option {
 
 // WithSessionPoolKeepAliveTimeout set timeout of keep alive requests for session in table.Client
 func WithSessionPoolKeepAliveTimeout(keepAliveTimeout time.Duration) Option {
-	return func(ctx context.Context, c *connection) error { return nil }
+	return func(ctx context.Context, c *Driver) error { return nil }
 }
 
 // WithSessionPoolCreateSessionTimeout set timeout for new session creation process in table.Client
 func WithSessionPoolCreateSessionTimeout(createSessionTimeout time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, tableConfig.WithCreateSessionTimeout(createSessionTimeout))
 		return nil
 	}
@@ -387,7 +387,7 @@ func WithSessionPoolCreateSessionTimeout(createSessionTimeout time.Duration) Opt
 
 // WithSessionPoolDeleteTimeout set timeout to gracefully close deleting session in table.Client
 func WithSessionPoolDeleteTimeout(deleteTimeout time.Duration) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, tableConfig.WithDeleteTimeout(deleteTimeout))
 		return nil
 	}
@@ -395,7 +395,7 @@ func WithSessionPoolDeleteTimeout(deleteTimeout time.Duration) Option {
 
 // WithIgnoreTruncated disables errors on truncated flag
 func WithIgnoreTruncated() Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(c.tableOptions, tableConfig.WithIgnoreTruncated())
 		return nil
 	}
@@ -406,7 +406,7 @@ func WithIgnoreTruncated() Option {
 // (before `WithTrace{Driver,Table,Scheme,Scripting,Coordination,Ratelimiter}` and other options)
 // If not defined - panic would not intercept with driver
 func WithPanicCallback(panicCallback func(e interface{})) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.panicCallback = panicCallback
 		c.options = append(c.options, config.WithPanicCallback(panicCallback))
 		return nil
@@ -415,7 +415,7 @@ func WithPanicCallback(panicCallback func(e interface{})) Option {
 
 // WithTraceTable returns table trace option
 func WithTraceTable(t trace.Table, opts ...trace.TableComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.tableOptions = append(
 			c.tableOptions,
 			tableConfig.WithTrace(
@@ -434,7 +434,7 @@ func WithTraceTable(t trace.Table, opts ...trace.TableComposeOption) Option {
 
 // WithTraceScripting scripting trace option
 func WithTraceScripting(t trace.Scripting, opts ...trace.ScriptingComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.scriptingOptions = append(
 			c.scriptingOptions,
 			scriptingConfig.WithTrace(
@@ -453,7 +453,7 @@ func WithTraceScripting(t trace.Scripting, opts ...trace.ScriptingComposeOption)
 
 // WithTraceScheme returns scheme trace option
 func WithTraceScheme(t trace.Scheme, opts ...trace.SchemeComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.schemeOptions = append(
 			c.schemeOptions,
 			schemeConfig.WithTrace(
@@ -472,7 +472,7 @@ func WithTraceScheme(t trace.Scheme, opts ...trace.SchemeComposeOption) Option {
 
 // WithTraceCoordination returns coordination trace option
 func WithTraceCoordination(t trace.Coordination, opts ...trace.CoordinationComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.coordinationOptions = append(
 			c.coordinationOptions,
 			coordinationConfig.WithTrace(
@@ -491,7 +491,7 @@ func WithTraceCoordination(t trace.Coordination, opts ...trace.CoordinationCompo
 
 // WithTraceRatelimiter returns ratelimiter trace option
 func WithTraceRatelimiter(t trace.Ratelimiter, opts ...trace.RatelimiterComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.ratelimiterOptions = append(
 			c.ratelimiterOptions,
 			ratelimiterConfig.WithTrace(
@@ -510,7 +510,7 @@ func WithTraceRatelimiter(t trace.Ratelimiter, opts ...trace.RatelimiterComposeO
 
 // WithRatelimiterOptions returns reatelimiter option
 func WithRatelimiterOptions(opts ...ratelimiterConfig.Option) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.ratelimiterOptions = append(c.ratelimiterOptions, opts...)
 		return nil
 	}
@@ -518,7 +518,7 @@ func WithRatelimiterOptions(opts ...ratelimiterConfig.Option) Option {
 
 // WithTraceDiscovery adds configured discovery tracer to Connection
 func WithTraceDiscovery(t trace.Discovery, opts ...trace.DiscoveryComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.discoveryOptions = append(
 			c.discoveryOptions,
 			discoveryConfig.WithTrace(
@@ -537,7 +537,7 @@ func WithTraceDiscovery(t trace.Discovery, opts ...trace.DiscoveryComposeOption)
 
 // WithTraceTopic adds configured discovery tracer to Connection
 func WithTraceTopic(t trace.Topic, opts ...trace.TopicComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.topicOptions = append(
 			c.topicOptions,
 			topicoptions.WithTrace(
@@ -556,7 +556,7 @@ func WithTraceTopic(t trace.Topic, opts ...trace.TopicComposeOption) Option {
 
 // WithTraceDatabaseSQL adds configured discovery tracer to Connection
 func WithTraceDatabaseSQL(t trace.DatabaseSQL, opts ...trace.DatabaseSQLComposeOption) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.databaseSQLOptions = append(
 			c.databaseSQLOptions,
 			xsql.WithTrace(
@@ -575,15 +575,15 @@ func WithTraceDatabaseSQL(t trace.DatabaseSQL, opts ...trace.DatabaseSQLComposeO
 
 // Private technical options for correct copies processing
 
-func withOnClose(onClose func(c *connection)) Option {
-	return func(ctx context.Context, c *connection) error {
+func withOnClose(onClose func(c *Driver)) Option {
+	return func(ctx context.Context, c *Driver) error {
 		c.onClose = append(c.onClose, onClose)
 		return nil
 	}
 }
 
 func withConnPool(pool *conn.Pool) Option {
-	return func(ctx context.Context, c *connection) error {
+	return func(ctx context.Context, c *Driver) error {
 		c.pool = pool
 		return pool.Take(ctx)
 	}